Refactor logging and enable in myloader
